Work around lost connection issue in Deno netcode
Deno doesn't maintain its connection pools in a separate thread,
so it's unsafe to block the event loop for an extended period.

The specific issue is that Deno needs to do a tick after the event loop
is process any new events that happened, such as remote peers telling us
they walked away on the keepalive.

Otherwise Deno will use an already-bad socket. This results in
an immediate error on the fetch() promise:

    connection closed before message completed